refactor: 블랙잭 상수 클래스 추가
