+++ src/main/java/blackjack/domain/BlackjackConstants.java
@@ -1,0 +1,15 @@
+package blackjack.domain;
+
+public enum BlackjackConstants {
+    MAX_SCORE(21);
+
+    private final int symbol;
+
+    BlackjackConstants(int symbol) {
+        this.symbol = symbol;
+    }
+
+    public int getSymbol() {
+        return symbol;
+    }
+}
